requirements.txt
Django==1.7.3
MySQL-python==1.2.5
Pillow==2.7.0
argparse==1.2.1
django-ajaximage==0.2.7
django-appconf==0.6
django-bootstrap-toolkit==2.15.0
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-imagekit==3.2.5
django-password-reset==0.7
django-registration-redux==1.1
pbr==0.10.7
pilkit==1.1.12
six==1.9.0
sqlparse==0.1.14
stevedore==1.2.0
virtualenv==12.0.5
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
wsgiref==0.1.2
